v3: extract pure theme-scope helpers into themeScope.mjs
ThemeProviderHost.js holds the React-side seam; the pure helpers it
relies on (pickDensity, hasThemeContent, scopedSelector,
buildScopedDetailCss + the THEME_SCOPE_ATTRIBUTE constants) now live in
a sibling .mjs so Node tests under tests/runtime/* can import() them
directly without a webpack/jest harness. Same pattern as
resolveRegion.mjs / matchRoute.mjs / platformServices.mjs already use.

Pure functional move — ThemeProviderHost re-uses the helpers verbatim.
